feat(error handler): Add BadRequestError handlers

+++ backend-service-nodejs/services/tasks-service.js
@@ -1,5 +1,5 @@
 const prisma = require('../prisma/prisma');
-const { NotFoundError } = require('../utils/errors');
+const { NotFoundError, BadRequestError } = require('../utils/errors');
 
 async function getTaskList() {
     const tasks = await prisma.task.findMany();
@@ -7,6 +7,10 @@ async function getTaskList() {
 }
 
 async function getTaskById(taskId) {
+
+    if (parseInt(taskId) <= 0) {
+        throw new BadRequestError("Invalid task ID");
+    }
     
     const task = await prisma.task.findUnique({
         where: {
@@ -22,6 +26,11 @@ async function getTaskById(taskId) {
 }
 
 async function createTask(task) {
+
+    if (task.completed !== false || task.completed !== true) {
+        throw new BadRequestError("Invalid task status; completed must be true or false");
+    }
+
     const result = await prisma.task.create({
         data: {
             ...task
@@ -32,6 +41,11 @@ async function createTask(task) {
 }
 
 async function deleteTask(taskId) {
+
+    if (parseInt(taskId) <= 0) {
+        throw new BadRequestError("Invalid task ID");
+    }
+
     const results = await prisma.task.delete({
         where: {
             id: parseInt(taskId)
